Added toast to warn guest if they try to vote on a snippet to log in first

+++ src/app/result/voting.tsx
@@ -14,7 +14,7 @@ import { toast } from "@/components/ui/use-toast";
 import { catchError, cn } from "@/lib/utils";
 
 interface VotingProps {
-  userId: User["id"];
+  userId: User["id"] | undefined;
   snippetId: Snippet["id"];
   usersVote?: SnippetVote;
 }
@@ -33,26 +33,34 @@ export function Voting({ userId, snippetId, usersVote }: VotingProps) {
           disabled={isPending}
           onClick={() => {
             startTransition(async () => {
-              try {
-                if (usersVote?.type === "UP") {
-                  await deleteVoteAction({
-                    userId,
-                    snippetId,
-                  });
-                } else {
-                  await upvoteSnippetAction({
-                    userId,
-                    snippetId,
-                  });
-                  toast({
-                    title: "Success.",
-                    description:
-                      "Thanks for your feedback! We will consider it.",
-                    variant: "default",
-                  });
+              if (userId) {
+                try {
+                  if (usersVote?.type === "UP") {
+                    await deleteVoteAction({
+                      userId,
+                      snippetId,
+                    });
+                  } else {
+                    await upvoteSnippetAction({
+                      userId,
+                      snippetId,
+                    });
+                    toast({
+                      title: "Success.",
+                      description:
+                        "Thanks for your feedback! We will consider it.",
+                      variant: "default",
+                    });
+                  }
+                } catch (err) {
+                  catchError(err);
                 }
-              } catch (err) {
-                catchError(err);
+              } else {
+                toast({
+                  title: "Warning",
+                  description: "You should sign in first to vote.",
+                  variant: "middle",
+                });
               }
             });
           }}
@@ -69,26 +77,34 @@ export function Voting({ userId, snippetId, usersVote }: VotingProps) {
           disabled={isPending}
           onClick={() => {
             startTransition(async () => {
-              try {
-                if (usersVote?.type === "DOWN") {
-                  await deleteVoteAction({
-                    userId,
-                    snippetId,
-                  });
-                } else {
-                  await downvoteSnippetAction({
-                    userId,
-                    snippetId,
-                  });
-                  toast({
-                    title: "Success.",
-                    description:
+              if (userId) {
+                try {
+                  if (usersVote?.type === "DOWN") {
+                    await deleteVoteAction({
+                      userId,
+                      snippetId,
+                    });
+                  } else {
+                    await downvoteSnippetAction({
+                      userId,
+                      snippetId,
+                    });
+                    toast({
+                      title: "Success.",
+                      description:
                       "Thanks for your feedback! We will consider it.",
-                    variant: "default",
-                  });
+                      variant: "default",
+                    });
+                  }
+                } catch (err) {
+                  catchError(err);
                 }
-              } catch (err) {
-                catchError(err);
+              } else {
+                toast({
+                  title: "Warning",
+                  description: "You should sign in first to vote.",
+                  variant: "middle",
+                });
               }
             });
           }}
